updating posts works now
still need to change how it works but i made it so it works with how we now use readdtos and how we have owner ID instead of an Owner object. This fix is only for the web api and all the methods in it works. the blazor is still having issues with editing a existing post